// src/app/services/enums/menuitemurlnum .ts
export enum MenuItemUrl {
    CultureGuide = '/manha/guia-de-cultura',
    BrandManual = '/manha/manual-da-marca',
    Compass = '/manha/bussola',
    Tree = '/manha/arvore',
    Guardians = '/manha/guardioes',
    Magazine = '/manha/dez-anos',
    Colab = '/noite/colab',
    Us = '/noite/nos',
    Stamp = '/noite/selo',
    Network = '/noite/rede',
    Talks = '/noite/palestras',
    Gabw = 'https://gabw.cc/',
    Contact = '/noite/contato',
    Timeline = '/tarde/linha-do-tempo'
}